backend/requirements.txt
fastapi[all]
requests
mysql-connector-python
dotenv
serverless-wsgi